CMake: make it possible to disable a lot of found dependencies
In some use cases, it might be disable to easily disable use of a
detected library. So in the gdal_check_package() macro add a CAN_DISABLE
option, that adds a cmake_dependent_option(GDAL_USE_foo) for package
foo. And in CMakeLists.txt files, instead of testing HAVE_foo, use
GDAL_USE_foo instead.

This is a work in progress. Not all libraries have been converted.

Fix a few invalid driver dependencies in the process.